+++ lib/race_browser_screen/race_browser.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:marklin_bluetooth/firebase/race.dart';
 import 'package:marklin_bluetooth/firebase/race_reference.dart';
 import 'package:marklin_bluetooth/firebase/races.dart';
-import 'package:marklin_bluetooth/race_viewer.dart';
+import 'package:marklin_bluetooth/race_browser_screen/race_viewer.dart';
 import 'package:marklin_bluetooth/utils.dart';
 import 'package:marklin_bluetooth/widgets.dart';
